Optimize logged webhook trimming
We were setting truncated_at whenever we trimmed,
rather than only setting it on untruncated webhooks.

During the course of this fix it was also clear we weren't using
indices effectively when trimming. We had one index on inserted_at,
but this table can be very very large, so it led to many long index
scans that needed to eliminate rows with a filter.

This creates separate partial indices for each query
(which together cover the table space), ensuring minimal scans.

Tested with production to make sure the indices are used.