Support adding indices to index tables
We use CREATE INDEX CONCURRENTLY, which doesn't work
with partitioned tables.

Instead, we need to create the index concurrently on each partition,
and then attach them to the 'parent' index.

Massages dbadapter to support this better.

Add a structured `Partition` type,
to describe a partition (parent table, name, suffix).

Add helper to gather existing partitions.

Add helper to align partition names to the parent table name,
which is necessary after renaming a table with partitions
(may be automated in the future after renaming a table).