Fix autoscaler bug
In 6d00be465ea741a9460f915a2db0311fd6212cf1, I changed how we report
high latency queues to Sentry. There was a mistake causing an
undefined local (`kw`).

This wasn't caught in tests, because we used `scope&.set_tags(*kw)`.
But `scope` was nil because Sentry wasn't configured for tests,
so the `set_tags` call wasn't made and the invalid local
wasn't evaluated.

This wasn't reported to Sentry because it happened in a background
thread, which Sentry doesn't listen for errors in.

This change does the following:

- Fixes the bug
- Uses the new `on_unhandled_exception` argument to
  `Amigo::Autoscaler`, which means we will report on
  unhandled exceptions.
- Tests the fix and that we're reporting unhandled errors correctly.
- Moves Sentry test helpers into a spec helper,
  since full Sentry test configuring is nontrivial
  (need to make sure there is a current hub in order
  for scope to work, for example).