Speed up stale row deletion on partitioned tables
If the replicator is partitioned, we need to delete stale rows on partition separately.
We DELETE with a LIMIT in chunks, but when we run this on the main table, it'll run the query
on every partition BEFORE applying the limit. You'll see this manifest with speed,
but also the planner using a sequential scan for the delete, rather than hitting an index.
Instead, DELETE from each partition in chunks, which will use the indices, and apply the limit properly.