Front-signalwire channel support attachments
It turns out, doing this with a media replicator isn't really feasible.
Instead, when syncing to Front, find the parent signalwire message,
and then download the attachments before syncing.

The issues were around race conditions (both message and media
need to sync in order to import), and storing media files:
at the point the replicator needs to download the media,
the additional 'get media list' is not any more complex.
The alternative would be storing the binary media,
but this doesn't seem necessary or desirable.