Signalwire media sync, include front/signalwire attachments (#958)
* Fix bug when backfill state machines fail

We were potentially zeroing out all data,
but then asking for one specific piece.
Instead, we need to be more deliberate, and recalculate
the entire state machine (after clearing out what is potentially
incorrect, which could be anything).

This fixes this bug across all files by removing
the http status-specific error messages,
and moving those messages to be handled explicitly
when calculating the state machine fails.

* Add Signalwire Media replicator

It's a child of the Signalwire Message replicator,
and uses that table to figure out what media needs to be synced.

* Front-signalwire channel support attachments

It turns out, doing this with a media replicator isn't really feasible.
Instead, when syncing to Front, find the parent signalwire message,
and then download the attachments before syncing.

The issues were around race conditions (both message and media
need to sync in order to import), and storing media files:
at the point the replicator needs to download the media,
the additional 'get media list' is not any more complex.
The alternative would be storing the binary media,
but this doesn't seem necessary or desirable.

* Bump actions/cache

* Support disabling async events for models, needed for database documents which may have binary content.

* Add helper so spec helpers can accumulate metadata. Use it for `truncate:`, so spec groups can truncate multiple tables.

* Database documents can be scheduled for deletion. Use for creating accessible URLs for Signalwire to download.