Advisory lock: Handle table oids > MAX_INT
Table OIDs are unsigned integers. If they are used as advisory
lock keys, we need to convert them to integers.

Any OID between MAX_INT and MAX_UINT should be biased to MIN_INT to 0.
We do not use a straight int->uint conversion, because we do not want
to change the advisory locks of existing working code.