Add integration-specific job semaphores
We need to migrate from a non-partitioned to a partitioned google
calendar events table, which is billions of rows (or very large,
in any case).

If we just allow this to start backfilling, it will saturate all
the workers for the org, and also consume a ton of DB CPU
(though everything works). To prevent this from happening,
we can semaphore the integration in particular.