Dynamically load tasks; sync target improvements; StopIteration fix (#963)
Improve sync target stat tracking

Track full call duration, not just remote calls,
to find slow queries.

Better cursor page name.

Use better summary names and calculations.

---

Support disabling sync targets

---

Expose str2inthash method and converter

So that replicators can convert a column into a hash value
for partitioning, add a converter.

This is useful when a 'parent' (like a calendar) needs to
store the hash used for children (events),
so clients can easily query.

---

Dynamically load tasks

Allow other modules to define tasks.

---

 Replace httpx and down to fix `StopIteration` bugs

The StopIteration problem with `down` and `httpx` turned
into a giant can of worms that I wasn't able to fix.
There was no way to tell the difference between redirects and 304s,
since we'd get a StopIteration for both.

To fix this, I decided to rip out Down, since it wasn't providing
useful abstraction over anything.
I also needed to rip out HTTPX, since its streaming implementation
was busted and wasn't easy to patch or replace.

Instead, we use the more mature, and in my opinion clearer,
`HTTP` gem. We wrap it more thoroughly with `chunked_download`.

This required some small changes to `Webhookdb::Http` to make
it more flexible for the `HTTP` responses (not just HTTParty).

It included a much more annoying change to handle the content type
and charset detection fixes we made on top of HTTPX and Down,
to handle the bad HTTP servers we deal with. These are small and
subtle things, so it's good they were tested!

Unfortunately the Webhook HTTP lib wrapper makes too many
assumptions, and we can't use it to unit test these charset changes.
So we use `thin` to start a webserver we can use
for better http testing.

Overall I think this code is much simpler and will be easier
to fix and modify. Look at all the exception types we had to
handle, for example- we'll need to add more, almost for sure,
but it's much more clear.