Fix connection errors during body reads
For a long time, we've been plagued by errors in Sentry due to
errors when reading an HTTP body (not on the initial connection).

Now that we have an in-process http server for testing,
I was able to reproduce this problem using some rack hijacking.
This also meant removing Thin for Puma (thin doesn't support
rack.hijack).

The fix was sort of obvious- wrapping the body-reading portion
of the code in the same sort of exception handling
that the initial connection uses.

Refactor the in-process http server into a spec helper module
so it can be reused elsewhere.