Add helper to align index names
When a service integration opaque id changes,
it is nice to be able to align index names to reflect
the new opaque id.

This is not necessary, as the routines now ignore differences
in opaque ids when searching for existing indices,
but it is nice to not have old names standing around.