Partition: No LIMIT when deleting stale rows
This can be very slow and result in many redundant seqscans
of a single table. Instead, either take the seqscan hit
on the partition once, or ideally, use the dedicated stale row
idx, which will delete extremely quickly
because we'll no longer have the subselect.

Note that reframing the LIMIT query USING and CTEs
did not solve the seqscan problem.