Ruby 3.4, Sidekiq 7+, Rack 3 (#973)
This is a significant upgrade change for Ruby 3.4, Rack 3, and Sidekiq 7/8. It does not require any migrations or downtime, but custom replicators may be impacted.

The biggest change are the Sidekiq and Rack updates, which are not backwards compatible.

Other changes are:

- No longer uses `redis` gem
- All header usage within replicators must be lowercased. Callouts to APIs can still be uppercased, but accessing any headers on webhook requests or responses must be lowercase.
- Remove `_nil_or_json` helper as it wasn't used/tested (it was for a legacy enterprise replicator)


* Update all outdated gems

* Sidekiq and Rack changes

* More test fixes

* update rubocop

* Rubocop changes

* Fix header case, patch rack ssl enforcer

* Fix decode cookie for rack 3.1

* update docker and rubyversion to 3.4.4

* Add test coverage, fix bugs in ViewApi

- Add coverage for Redis#fetch_url
- Remove _nil_or_json, not used (used in enterprise Plaid)
- Add coverage to ViewApi endpoints

The ViewApi endpoints weren't covered, and there were several bugs.
It's all covered by tests and the known bugs are fixed.

* Use lowercase headers in all app code

Leave them uppercase in test code to help test case insensitivity.

* Add rake tests to retry all jobs in the dead and retry set

* fix tests casing broke