Fix default max connections in web workers
The max threads in the pool was set to WEB_CONCURRENCY,
which is the number of processes, NOT the number of threads.
We should use RAILS_MAX_THREADS as the pool size instead.

Also add tests for dbutil configuration.

Should fix issues with slow web response times and pool timeouts.