Use timeouts when calling icalendar urls
We were calling icalendar urls (usually icalproxy) without a timeout.
This was an oversight, possibly removed when the chunked download
was rewritten a while ago.
This could result in hung clients.
See https://github.com/webhookdb/icalproxy/pull/6

Ideally we will get errors reported if there are server problems,
rather than a bunch of hung workers.

It is of course still possible that we'll get hung workers,
but hopefully not!