Fix race condition with windows timezones
We load the windows timezone database on demand.
But we weren't using locks, meaning that,
while the database was initializing, other threads could
misread it, or even mutate the underlying mapping.

Instead, use a mutex with double-check locking
(for a fast path) so the database is only loaded once per process
and it isn't thrashed.

This seems to fix the timezone database errors.